move embedded-hal_02 to subdir for I2C
