+++ src/i2c.rs
@@ -1,5 +1,4 @@
 use core::ops::Deref;
-use embedded_hal::blocking::i2c::{Read, Write, WriteRead};
 
 use crate::pac::i2c1;
 use crate::rcc::{Enable, Reset};
@@ -17,6 +16,8 @@ use crate::gpio::{gpioa, gpiob, gpioc, gpioh};
 
 use crate::rcc::Clocks;
 use crate::time::{Hertz, U32Ext};
+
+mod hal_02;
 
 #[derive(Debug, Eq, PartialEq)]
 pub enum DutyCycle {
@@ -264,7 +265,7 @@ where
 }
 
 trait I2cCommon {
-    fn write_bytes(&mut self, addr: u8, bytes: &[u8]) -> Result<(), Error>;
+    fn write_bytes(&mut self, addr: u8, bytes: impl Iterator<Item = u8>) -> Result<(), Error>;
 
     fn send_byte(&self, byte: u8) -> Result<(), Error>;
 
@@ -275,7 +276,7 @@ impl<I2C, PINS> I2cCommon for I2c<I2C, PINS>
 where
     I2C: Instance,
 {
-    fn write_bytes(&mut self, addr: u8, bytes: &[u8]) -> Result<(), Error> {
+    fn write_bytes(&mut self, addr: u8, bytes: impl Iterator<Item = u8>) -> Result<(), Error> {
         // Send a START condition
         self.i2c.cr1.modify(|_, w| w.start().set_bit());
 
@@ -313,7 +314,7 @@ where
 
         // Send bytes
         for c in bytes {
-            self.send_byte(*c)?;
+            self.send_byte(c)?;
         }
 
         // Fallthrough is success
@@ -347,103 +348,5 @@ where
 
         let value = self.i2c.dr.read().bits() as u8;
         Ok(value)
-    }
-}
-
-impl<I2C, PINS> WriteRead for I2c<I2C, PINS>
-where
-    I2C: Instance,
-{
-    type Error = Error;
-
-    fn write_read(&mut self, addr: u8, bytes: &[u8], buffer: &mut [u8]) -> Result<(), Self::Error> {
-        self.write_bytes(addr, bytes)?;
-        self.read(addr, buffer)?;
-
-        Ok(())
-    }
-}
-
-impl<I2C, PINS> Write for I2c<I2C, PINS>
-where
-    I2C: Instance,
-{
-    type Error = Error;
-
-    fn write(&mut self, addr: u8, bytes: &[u8]) -> Result<(), Self::Error> {
-        self.write_bytes(addr, bytes)?;
-
-        // Send a STOP condition
-        self.i2c.cr1.modify(|_, w| w.stop().set_bit());
-
-        // Wait for STOP condition to transmit.
-        while self.i2c.cr1.read().stop().bit_is_set() {}
-
-        // Fallthrough is success
-        Ok(())
-    }
-}
-
-impl<I2C, PINS> Read for I2c<I2C, PINS>
-where
-    I2C: Instance,
-{
-    type Error = Error;
-
-    fn read(&mut self, addr: u8, buffer: &mut [u8]) -> Result<(), Self::Error> {
-        if let Some((last, buffer)) = buffer.split_last_mut() {
-            // Send a START condition and set ACK bit
-            self.i2c
-                .cr1
-                .modify(|_, w| w.start().set_bit().ack().set_bit());
-
-            // Wait until START condition was generated
-            while self.i2c.sr1.read().sb().bit_is_clear() {}
-
-            // Also wait until signalled we're master and everything is waiting for us
-            loop {
-                let sr2 = self.i2c.sr2.read();
-                if !(sr2.msl().bit_is_clear() && sr2.busy().bit_is_clear()) {
-                    break;
-                }
-            }
-
-            // Set up current address, we're trying to talk to
-            self.i2c
-                .dr
-                .write(|w| unsafe { w.bits((u32::from(addr) << 1) + 1) });
-
-            // Wait until address was sent
-            loop {
-                self.check_and_clear_error_flags()?;
-                if self.i2c.sr1.read().addr().bit_is_set() {
-                    break;
-                }
-            }
-
-            // Clear condition by reading SR2
-            self.i2c.sr2.read();
-
-            // Receive bytes into buffer
-            for c in buffer {
-                *c = self.recv_byte()?;
-            }
-
-            // Prepare to send NACK then STOP after next byte
-            self.i2c
-                .cr1
-                .modify(|_, w| w.ack().clear_bit().stop().set_bit());
-
-            // Receive last byte
-            *last = self.recv_byte()?;
-
-            // Wait for the STOP to be sent.
-            while self.i2c.cr1.read().stop().bit_is_set() {}
-
-            // Fallthrough is success
-            Ok(())
-        } else {
-            Err(Error::OVERRUN)
-        }
     }
 }
